Homework2/project/requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pillow>=9.5
torch>=2.3
torchvision>=0.18
